tools/datasets/requirements.txt
datasets
numpy
onnx
huggingface_hub
Pillow
opencv-python-headless
onnxruntime
transformers
optimum
